fix: keep styling in single files
