+++ packages/components/card/src/asset-card/AssetCard.styles.ts
@@ -1,4 +1,5 @@
 import { css } from 'emotion';
+import type { ObjectInterpolation } from 'emotion';
 import tokens from '@contentful/f36-tokens';
 
 export const getAssetCardStyles = () => {
@@ -36,10 +37,27 @@ export const getAssetCardStyles = () => {
       paddingRight: tokens.spacingXs,
       paddingTop: 0,
     }),
-    root: css({
-      borderRadius: tokens.borderRadiusMedium,
-      padding: 0,
-      textAlign: 'center',
-    }),
+    root: ({ isSelected }) => {
+      const styles: ObjectInterpolation<undefined> = {
+        borderRadius: tokens.borderRadiusMedium,
+        padding: 0,
+        textAlign: 'center',
+      };
+
+      if (isSelected) {
+        styles.backgroundColor = tokens.colorWhite;
+        styles.borderColor = tokens.blue500;
+        styles.boxShadow = `0 0 0 1px ${tokens.gray300}`;
+      }
+
+      return css(styles);
+    },
+    sizeWrapper: ({ size }) =>
+      css({
+        height:
+          size === 'small'
+            ? `calc(9px * ${tokens.fontBaseDefault})`
+            : `calc(13px * ${tokens.fontBaseDefault})`,
+      }),
   };
 };
