+++ packages/components/card/src/card/Card.styles.ts
@@ -8,14 +8,17 @@ export const getCardStyles = () => {
       paddingLeft: tokens.spacingL,
       paddingRight: tokens.spacingL,
     }),
-    header: css({
-      gridColumn: 'content',
-      gridRow: 'header',
-    }),
-    headerWithActions: css({
-      alignItems: 'flex-end',
-      paddingRight: tokens.spacingXs,
-      paddingTop: tokens.spacingXs,
-    }),
+    header: ({ padding }) =>
+      css({
+        gridColumn: 'content',
+        gridRow: 'header',
+        padding: padding === 'large' ? tokens.spacingL : tokens.spacingM,
+      }),
+    headerWithActions: ({ padding }) =>
+      css({
+        alignItems: 'flex-end',
+        paddingRight: tokens.spacingXs,
+        paddingTop: padding === 'large' ? tokens.spacingM : tokens.spacingXs,
+      }),
   };
 };
