+++ packages/components/card/src/inline-entry-card/InlineEntryCard.tsx
@@ -1,34 +1,11 @@
 import React from 'react';
 import { css, cx } from 'emotion';
-import tokens from '@contentful/f36-tokens';
-import { getEntityStatusStyles } from '@contentful/f36-core';
 import { Text } from '@contentful/f36-typography';
 
 import { BaseCard } from '../base-card/BaseCard';
 import type { EntryCardInternalProps } from '../entry-card/EntryCard';
 import { getInlineEntryCardStyles } from './InlineEntryCard.styles';
 import { CardActions } from '../base-card/CardActions';
-
-const generateStyles = ({
-  status,
-}: {
-  status: InlineEntryCardProps['status'];
-}) => {
-  const statusColors = getEntityStatusStyles({ status });
-
-  return css({
-    '&::before': {
-      backgroundColor: statusColors.color,
-      bottom: 0,
-      content: '""',
-      display: 'block',
-      left: 0,
-      position: 'absolute',
-      top: 0,
-      width: tokens.spacing2Xs,
-    },
-  });
-};
 
 export type InlineEntryCardInternalProps = Omit<
   EntryCardInternalProps,
@@ -50,12 +27,11 @@ export const InlineEntryCard = ({
       {actions}
     </CardActions>
   );
-  const statusStyles = generateStyles({ status });
 
   return (
     <BaseCard
       {...otherProps}
-      className={cx(styles.root, statusStyles, className)}
+      className={cx(styles.root({ status }), className)}
       contentBodyProps={{ className: css({ padding: 0 }) }}
       header={header}
     >
